+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Unless otherwise specified, any version comparison below is the comparison of th
 
 - (API) Fixed method `Container.removeAllViewers()` won't remove all viewers from the container.
 - `RAK_SEND_COOKIE` is set to `true` by default in `AllayNetworkInterface` to resolve connection issue.
+- Fixed a bug where calling method `Container.removeViewer()` server-side will make the client no longer able to open any container.
 - Fixed sneaking state resolution while flying to prevent incorrect interaction behavior.
 - Fixed mobile block interaction regression by always attempting item use on block clicks.